Changed sql config files and made changes to db tables

+++ data_handler.js
@@ -47,7 +47,7 @@ knex.schema.createTableIfNotExists('parts', function(table){
   table.string('attribute');
   table.string('description');
   table.integer('recamt');
-  table.integer('price');
+  table.decimal('price');
   table.integer('quantity');
